Finished LoG

+++ Project 2/lin_img_conv.m
@@ -2,8 +2,8 @@ function O = lin_img_conv(I,F)
 %LIN_IMG_CONV Does the same thing as imfilter()
 %   Does the same thing as imfilter()
     O = zeros(size(I));
-    F = int16(F);
-    I = int16(I);
+    F = double(F);
+    I = double(I);
     % Determine what row/col to start, use zero padding
     eff_size = floor(size(F,1)/2);
     % WHY DOES MATLAB HAVE TO INDEX AT 1?!?!?   AHHAHHAHAHAIWHBVYIRBVUYAYC
